# Dramatically speed up how shortcuts are found, using cells to restrict