--- a/backend/src/shortcuts.rs
+++ b/backend/src/shortcuts.rs
@@ -1,8 +1,8 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 
 use crate::map_model::Direction;
 use crate::route::Router;
-use crate::{MapModel, Neighbourhood, RoadID};
+use crate::{Cell, MapModel, Neighbourhood, RoadID};
 use geo::{Euclidean, Length, LineString};
 use geojson::Feature;
 
@@ -18,107 +18,113 @@ pub struct Path {
 }
 
 impl Shortcuts {
-    pub fn new(map: &MapModel, neighbourhood: &Neighbourhood) -> Self {
+    pub fn new(map: &MapModel, neighbourhood: &Neighbourhood, cells: &Vec<Cell>) -> Self {
         let router_input_after = neighbourhood.shortcuts_router_input_after(map);
         let router_input_before = neighbourhood.shortcuts_router_input_before(map);
         let router_after = Router::new(&router_input_after, 1.0);
+
         let mut paths = Vec::new();
         let mut count_per_road = HashMap::new();
-        let mut checked_routes: HashSet<(RoadID, RoadID)> = HashSet::new();
 
-        for start_i in &neighbourhood.border_intersections {
-            let start_intersection = map.get_i(*start_i);
-            for start_r in &start_intersection.roads {
-                // It's not a "shortcut" unless it starts outside the interior and cuts through the interior.
-                if !neighbourhood.main_roads.contains(start_r) {
-                    continue;
-                }
-                for end_i in &neighbourhood.border_intersections {
-                    if start_i == end_i {
+        // A shortcut is contained within a cell, because cells can't cross modal filters or main
+        // roads, and neither can shortcuts. Only looking for candidate shortcuts in each cell is
+        // much faster than using all pairs of the neighbourhood's border intersections, because
+        // all the pairs belonging to different cells will need to go along or cross a main road to
+        // achieve a route.
+        for cell in cells {
+            for start_i in &cell.border_intersections {
+                let start_intersection = map.get_i(*start_i);
+                for start_r in &start_intersection.roads {
+                    // It's not a "shortcut" unless it starts outside the interior and cuts through
+                    // the interior.
+                    if !neighbourhood.main_roads.contains(start_r) {
                         continue;
                     }
-                    let end_intersection = map.get_i(*end_i);
-                    'next_road: for end_r in &end_intersection.roads {
-                        // It's not a "shortcut" unless it ends outside the interior after cutting through.
-                        if !neighbourhood.main_roads.contains(end_r) {
+                    for end_i in &cell.border_intersections {
+                        if start_i == end_i {
                             continue;
                         }
+                        let end_intersection = map.get_i(*end_i);
+                        'next_road: for end_r in &end_intersection.roads {
+                            // It's not a "shortcut" unless it ends outside the interior after
+                            // cutting through.
+                            if !neighbourhood.main_roads.contains(end_r) {
+                                continue;
+                            }
 
-                        // It's possible to repeat the route request when the same road is
-                        // connected to multiple border intersections
-                        if checked_routes.contains(&(*start_r, *end_r)) {
-                            continue;
-                        }
-                        checked_routes.insert((*start_r, *end_r));
-
-                        // TODO We could get a little more precision by starting from the correct
-                        // end of the road, but it doesn't matter
-                        let Some(route) =
-                            router_after.route_from_roads(&router_input_after, *start_r, *end_r)
-                        else {
-                            continue;
-                        };
-
-                        let mut shortcut_length = 0.0;
-
-                        let interior_steps = {
-                            let mut steps = route.steps.iter();
-                            let first_step = steps.next().expect("route can't be empty");
-                            let first_road = map.get_r(first_step.0);
-                            shortcut_length += Euclidean.length(&first_road.linestring);
-                            let mut steps_reversed = steps.rev();
-                            let Some(final_step) = steps_reversed.next() else {
+                            // TODO We could get a little more precision by starting from the
+                            // correct end of the road, but it doesn't matter
+                            let Some(route) = router_after.route_from_roads(
+                                &router_input_after,
+                                *start_r,
+                                *end_r,
+                            ) else {
                                 continue;
                             };
-                            let final_road = map.get_r(final_step.0);
-                            shortcut_length += Euclidean.length(&final_road.linestring);
 
-                            // re-reverse back to the original ordering, but without first and final steps
-                            steps_reversed.rev()
-                        };
+                            let mut shortcut_length = 0.0;
 
-                        let mut shortcut_roads = Vec::new();
-                        for (r, _direction) in interior_steps {
-                            // For the purpose of counting unique shortcuts, only the first and
-                            // final steps should be on main roads.
-                            //
-                            // If we've left the interior, only the portion inside the interior counts
-                            // as distinct plausible shortcut.
-                            //
-                            // If a route leaves and re-enters the interior through another border
-                            // intersection, then that route will be counted as two distinct
-                            // shortcuts.
-                            if neighbourhood.main_roads.contains(&r) {
-                                break 'next_road;
+                            let interior_steps = {
+                                let mut steps = route.steps.iter();
+                                let first_step = steps.next().expect("route can't be empty");
+                                let first_road = map.get_r(first_step.0);
+                                shortcut_length += Euclidean.length(&first_road.linestring);
+                                let mut steps_reversed = steps.rev();
+                                let Some(final_step) = steps_reversed.next() else {
+                                    continue;
+                                };
+                                let final_road = map.get_r(final_step.0);
+                                shortcut_length += Euclidean.length(&final_road.linestring);
+
+                                // re-reverse back to the original ordering, but without first and final steps
+                                steps_reversed.rev()
+                            };
+
+                            let mut shortcut_roads = Vec::new();
+                            for (r, _direction) in interior_steps {
+                                // For the purpose of counting unique shortcuts, only the first and
+                                // final steps should be on main roads.
+                                //
+                                // If we've left the interior, only the portion inside the interior
+                                // counts as distinct plausible shortcut.
+                                //
+                                // If a route leaves and re-enters the interior through another
+                                // border intersection, then that route will be counted as two
+                                // distinct shortcuts.
+                                if neighbourhood.main_roads.contains(&r) {
+                                    break 'next_road;
+                                }
+
+                                shortcut_roads.push(*r);
                             }
 
-                            shortcut_roads.push(*r);
-                        }
-
-                        for r in shortcut_roads {
-                            let road = map.get_r(r);
-                            *count_per_road.entry(road.id).or_insert(0) += 1;
-                            shortcut_length += Euclidean.length(&road.linestring);
-                        }
-
-                        // How long is the shortest route through the original router, using this
-                        // neighbourhood or not?
-                        let direct_length = match map.router_before.route_from_roads(
-                            &router_input_before,
-                            *start_r,
-                            *end_r,
-                        ) {
-                            Some(route) => Euclidean.length(&route.to_linestring(map)),
-                            None => {
-                                warn!("Found a shortcut from {start_r} to {end_r}, but not a route using the whole map");
-                                shortcut_length
+                            // Only increase count_per_road after verifying above that the shortcut
+                            // is indeed valid and doesn't cross any main roads
+                            for r in shortcut_roads {
+                                let road = map.get_r(r);
+                                *count_per_road.entry(road.id).or_insert(0) += 1;
+                                shortcut_length += Euclidean.length(&road.linestring);
                             }
-                        };
-                        let directness = shortcut_length / direct_length;
-                        paths.push(Path {
-                            steps: route.steps,
-                            directness,
-                        });
+
+                            // How long is the shortest route through the original router, using this
+                            // neighbourhood or not?
+                            let direct_length = match map.router_before.route_from_roads(
+                                &router_input_before,
+                                *start_r,
+                                *end_r,
+                            ) {
+                                Some(route) => Euclidean.length(&route.to_linestring(map)),
+                                None => {
+                                    warn!("Found a shortcut from {start_r} to {end_r}, but not a route using the whole map");
+                                    shortcut_length
+                                }
+                            };
+                            let directness = shortcut_length / direct_length;
+                            paths.push(Path {
+                                steps: route.steps,
+                                directness,
+                            });
+                        }
                     }
                 }
             }