candidates

FIXES #124 and makes huge progress on #345.

Previously, we used every pair of border intersections in the neighbourhood as candidates for shortcuts. This was wasteful, because if we try to find a route between different cells, then by the definition of cells, the route has to cross or go along a main road. In one example larger neighbourhood, there were 145,000 wasted routes calculated, for only 1,000 valid ones. Calculating cells is cheap and we do it anyway, so we can simplify the shortcut calculation and only look for shortcuts within one cell at a time.

#354 proposed a check for border intersections at every intersection along a candidate route. We don't need this here, because we're never trying to cross over a main road now with the requests in the first place.

There are two test diffs. In Bristol, the shortcuts before this PR look like this:
![image](https://github.com/user-attachments/assets/7c97e44a-7489-46d8-9721-f8a03ae1b2a1)
With this PR, we lose the shortcuts on Victoria Avenue:
![image](https://github.com/user-attachments/assets/a2cddf82-2347-4d69-9d38-9bfe01177663)
This is caused by a funny small cell on the western edge:
![image](https://github.com/user-attachments/assets/ef4322c2-9b23-4527-833b-2405a8f5f169)
I think this is acceptable, because this is a case where the automatic main road detection is a little wrong. When a one-way main road points into a neighbourhood like this, it feels to me more like an interior road; the border should be a bit further back. If I manually reclassify and make it look correct:
![image](https://github.com/user-attachments/assets/ca786eae-1a68-4008-a50f-b51b58b7393e)
Then the shortcut on Victoria Avenue works fine:
![image](https://github.com/user-attachments/assets/643e7f94-cf16-4796-aedc-6848f31680c0)

Strasbourg has a change only to the number of shortcuts along the eastern edge. Two shortcuts get perfectly repeated; they are totally identical. One is:
![image](https://github.com/user-attachments/assets/1f2a81b4-0c98-49ed-8b2e-ed5149e8b47b)
The other is:
![image](https://github.com/user-attachments/assets/7974b3f8-76d8-4851-9e70-968689109635)
I'm not entirely sure why, but the eastern edge where both of these duplicates start looks odd to me, with an interior road in between two main roads. I'm not sure best practice is for neighbourhood definitions to include both of the parallel dual carriageway pieces, so we could even consider editing this test definition. (Maybe the same for Bristol too?)
![image](https://github.com/user-attachments/assets/c92dab30-2d12-4756-9ac1-d7f3842ca74f)

In terms of speed impact, the benchmark even for the pretty simple test areas gets much better, because still pruning by cells is useful, even if there's only "one" main road loop.
```
shortcuts in bristol_east
                        time:   [1.9851 ms 2.1903 ms 2.4327 ms]
                        change: [-28.750% -24.053% -19.337%] (p = 0.00 < 0.05)
                        Performance has improved.

shortcuts in bristol_west
                        time:   [6.0424 ms 6.2505 ms 6.5135 ms]
                        change: [-50.039% -47.191% -44.010%] (p = 0.00 < 0.05)
                        Performance has improved.

shortcuts in strasbourg time:   [2.5492 ms 2.5572 ms 2.5660 ms]
                        change: [-33.187% -31.986% -30.802%] (p = 0.00 < 0.05)
                        Performance has improved.
```

And for the huge city centre areas, I haven't measured in release mode yet, but just on dev mode, one of my examples goes from 3-4s down to 200ms! I'll try release mode on the #345 example, but I bet that 6.2s example is subsecond now.