@dabreegster 2025-05-08
![image](https://github.com/user-attachments/assets/1fc43b8d-4a1d-4d1e-ad48-745e742f288f)
In this massive area in release mode, total update time is down to 251ms: Finding cells took 3ms, rendering cells took 166ms, finding shortcuts took 82ms