@michaelkirk 2025-05-08
> With this PR, we lose the shortcuts on Victoria Avenue:
>
> This is caused by a funny small cell on the western edge:

I agree that it's a strange cell, and I agree that it's acceptable that the cell exists given the input data.

That said, I don't really understand why this change would result in a big swing in shortcuts for this case.